Creates Student & StudentCourse Models
Creates both models with its proper validations and setups new associations.
A Student belongs to a School and is able to have as many StudentCourses as it needs.
A StudentCourse belongs to both Courses and Students, also a Course can have as many StudentCourse as it needs, letting this model play a role as a join model.